+++ app/routes/events.tsx
@@ -1,7 +1,8 @@
 import { redirect, type LoaderFunction } from '@remix-run/node';
-import { Link, Outlet, useLoaderData, useLocation } from '@remix-run/react';
+import { useLoaderData } from '@remix-run/react';
 import { getUserId } from '~/server/auth.server';
 import { getEvents } from '~/server/event.server';
+import Sidebar from '~/shared/components/Sidebar';
 
 export const loader: LoaderFunction = async ({ request }) => {
   const userId = getUserId(request);
@@ -13,37 +14,13 @@ export const loader: LoaderFunction = async ({ request }) => {
 
 export default function EventList() {
   const data: any = useLoaderData();
-  const location = useLocation();
   const events = data?.event;
 
   return (
     <>
       <div className="home">
         <div className="row">
-          <div className={`col col-3 sidebar`}>
-            <div className="sidebar-header">
-              <i className="icon icon-list"></i>
-              {location.pathname === '/events' ? (
-                <Link to={'/events/create'}>
-                  <i className="icon icon-plus"></i>
-                </Link>
-              ) : (
-                <Link to={'/events'}>
-                  <i className="icon icon-arrow-left"></i>
-                </Link>
-              )}
-            </div>
-            {location.pathname === '/events' ? (
-              events.map((event: any) => (
-                <div key={event.id}>
-                  <Link to={`/events/${event.id}/edit`}>{event.title}</Link>
-                </div>
-              ))
-            ) : (
-              <Outlet />
-            )}
-          </div>
-
+          <Sidebar events={events} />
           <div className="col col-9"></div>
         </div>
       </div>
